Renamed api service to gateway
