+++ internal/gateway/handler.go
@@ -1,4 +1,4 @@
-package api
+package gateway
 
 import (
 	"net/http"
